+++ packages/executor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # fuzd-executor
 
+## 0.9.6
+
+### Patch Changes
+
+- fixes + allow more timing options
+
 ## 0.9.5
 
 ### Patch Changes